Added more Controllers

+++ src/main/java/model/Client.java
@@ -17,7 +17,7 @@ public class Client {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int personalId;
+    private Long personalId;
     private String firstName;
     private String lastName;
     private String age;
